Added a system to save users data
Now, when a user runs the .exe program from the Code, the system I'll be able to save data inserted by the user, using the keyboard.

Now the system I'll be able to save: name, age, sex, and personal register (ID). Save the single answer, one by one. The data can be used in some ways.